requirements.txt
Flask==2.0.1
Flask-WTF==1.0.0
requests==2.26.0
beautifulsoup4==4.9.3
python-whois==0.7.3
dnspython==2.1.0
tldextract==3.1.2
python-Levenshtein==0.12.2
nltk==3.6.3
schedule==1.1.0
networkx==2.6.3
Flask-Limiter==3.5.0
Flask-Caching==2.0.2
Flask-SocketIO==5.3.2
python-SocketIO==5.5.0
python-dotenv==0.19.0
gunicorn==20.1.0
yara-python==4.2.3
websockets==10.3
aiohttp==3.8.1
geopy==2.2.0
werkzeug==2.0.3 